requirements.txt
# Python dependencies for SPECTRA-Lab Platform
# Generated for compatibility with existing codebase

# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
scipy>=1.11.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0

# Instrument control
pyvisa>=1.12.0
pyvisa-py>=0.5.0

# API & Web
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Version management
packaging>=23.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
ruff>=0.0.280
mypy>=1.4.0

# Machine Learning (Core)
scikit-learn>=1.3.2
joblib>=1.3.0

# Deep Learning & RLHF
torch>=2.1.0
transformers>=4.44.0
datasets>=2.20.0
accelerate>=0.34.0
trl>=0.9.6
peft>=0.11.1
evaluate>=0.4.2

# AutoML & Optimization
optuna>=3.3.0

# Explainability (SHAP & LIME)
shap>=0.42.0
lime>=0.2.0.1

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.16.0

# Utilities
rich>=13.7.0
jsonschema>=4.22.0
tqdm>=4.66.4
